+++ boomoseries-Amazon-api/boomoseries-Amazon-api/Controllers/AmazonController.cs
@@ -20,7 +20,7 @@ namespace AmazonPrime_Microservice.Controllers
         }
 
         [HttpGet("movies")]
-        public async Task<IActionResult> GetMovies(float? min_rating)
+        public async Task<IActionResult> GetMovies(double? min_rating)
         {
             var movies = dataContext.Watchables.Where(Watchable => Watchable.Type == "Movie");
             List<WatchableDTO> movieDTOs = new();
@@ -68,7 +68,7 @@ namespace AmazonPrime_Microservice.Controllers
             }
         }
         [HttpGet("series")]
-        public async Task<IActionResult> GetSeries(float? min_rating)
+        public async Task<IActionResult> GetSeries(double? min_rating)
         {
             var series = dataContext.Watchables.Where(Watchable => Watchable.Type == "TV Show");
             List<WatchableDTO> seriesDTOs = new();
