+++ boomoseries-IMDB-api/boomoseries-IMDB-api/Controllers/IMDBController.cs
@@ -20,7 +20,7 @@ namespace boomoseries_IMDB_api.Controllers
         }
 
         [HttpGet("movies")]
-        public async Task<IActionResult> GetMovies(float? min_rating)
+        public async Task<IActionResult> GetMovies(double? min_rating)
         {
             var movies = dataContext.Watchables.Where(watchable => watchable.Type == "Movie");
             List<WatchableDTO> movieDTOs = new();
@@ -69,7 +69,7 @@ namespace boomoseries_IMDB_api.Controllers
         }
 
         [HttpGet("series")]
-        public async Task<IActionResult> GetSeries(float? min_rating)
+        public async Task<IActionResult> GetSeries(double? min_rating)
         {
             var series = dataContext.Watchables.Where(watchable => watchable.Type == "TV Show");
             List<WatchableDTO> seriesDTOs = new();
